connect critical data to one file
